requirements.txt
beautifulsoup4
pyshorteners
requests
python-telegram-bot==20.0a0
python-dotenv
